refactor: Flatten HassApiClient into HomeAssistantApiClient
These basically did the same thing and the separation was confusing.
