+++ gradle.properties
@@ -13,7 +13,7 @@ kotlin.mpp.stability.nowarn=true
 
 GROUP=com.codellyrandom.hassle
 POM_ARTIFACT_ID=hassle
-VERSION_NAME=0.1.3
+VERSION_NAME=0.1.4
 
 POM_NAME=Hassle
 POM_DESCRIPTION=Interact with your Home Assistant server using Kotlin.
