+++ src/commonMain/kotlin/com/codellyrandom/hassle/WebSocketSession.kt
@@ -1,6 +1,5 @@
 package com.codellyrandom.hassle
 
-import co.touchlab.kermit.Kermit
 import com.codellyrandom.hassle.core.mapping.ObjectMapper
 import io.ktor.client.features.websocket.*
 import io.ktor.http.cio.websocket.*
@@ -9,8 +8,6 @@ internal class WebSocketSession(
     delegate: DefaultClientWebSocketSession,
     val objectMapper: ObjectMapper
 ) : ClientWebSocketSession by delegate {
-
-    private val logger = Kermit()
     suspend fun callWebSocketApi(message: String) = send(message)
 
     suspend inline fun <reified M : Any> callWebSocketApi(message: M) = send(objectMapper.toJson(message))
